Timeseries added and project directory cleaned up.
